pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "flood-qc-ops"
version = "0.1.0"
description = "Esqueleto operacional para hidrologia e previsao com foco em SQLite, Streamlit e artefatos locais."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
data = [
  "numpy>=1.26",
  "pandas>=2.1",
  "requests>=2.32",
]
geo = [
  "rasterio>=1.3",
]
ui = [
  "streamlit>=1.40",
  "plotly>=5.20",
  "folium>=0.18",
  "streamlit-folium>=0.24",
  "branca>=0.8",
  "affine>=2.4",
]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
packages = []

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"